[style #22] chattingroom 디렉토리 chatting으로 변경 및 리팩토링
