+++ src/main/java/com/freepath/devpath/chatting/command/application/dto/request/ChattingRoomCreateRequest.java
@@ -1,0 +1,7 @@
+package com.freepath.devpath.chatting.command.application.dto.request;
+
+public class ChattingRoomCreateRequest {
+    private int boardId;
+    private String chattingRoomTitle;
+    private int userCount;
+}
