+++ src/main/java/com/freepath/devpath/chatting/command/application/dto/response/ChattingRoomCommandResponse.java
@@ -1,0 +1,10 @@
+package com.freepath.devpath.chatting.command.application.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class ChattingRoomCommandResponse {
+    private int chattingRoomId;
+}
