+++ src/main/java/com/freepath/devpath/chatting/command/domain/aggregate/ChattingRole.java
@@ -1,0 +1,5 @@
+package com.freepath.devpath.chatting.command.domain.aggregate;
+
+public enum ChattingRole {
+    ONE,OWNER, MEMBER
+}
