--- src/main/java/com/freepath/devpath/chattingroom/command/application/dto/request/ChattingRoomCreateRequest.java
@@ -1,7 +1,0 @@
-package com.freepath.devpath.chattingroom.command.application.dto.request;
-
-public class ChattingRoomCreateRequest {
-    private int boardId;
-    private String chattingRoomTitle;
-    private int userCount;
-}
