--- src/main/java/com/freepath/devpath/chattingroom/command/application/dto/response/ChattingRoomCommandResponse.java
@@ -1,10 +1,0 @@
-package com.freepath.devpath.chattingroom.command.application.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ChattingRoomCommandResponse {
-    private int chattingRoomId;
-}
