--- src/main/java/com/freepath/devpath/chattingroom/command/domain/aggregate/ChattingRole.java
@@ -1,5 +1,0 @@
-package com.freepath.devpath.chattingroom.command.domain.aggregate;
-
-public enum ChattingRole {
-    ONE,OWNER, MEMBER
-}
