--- src/main/java/com/freepath/devpath/chattingroom/command/domain/repository/ChattingRoomRepository.java
@@ -1,8 +1,0 @@
-package com.freepath.devpath.chattingroom.command.domain.repository;
-
-import com.freepath.devpath.chattingroom.command.domain.aggregate.ChattingRoom;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ChattingRoomRepository extends JpaRepository<ChattingRoom, Integer> {
-
-}
