--- src/main/java/com/freepath/devpath/chattingroom/query/dto/ChattingRoomResponse.java
@@ -1,12 +1,0 @@
-package com.freepath.devpath.chattingroom.query.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-import java.util.List;
-
-@Getter
-@Builder
-public class ChattingRoomResponse {
-    private final List<ChattingRoomDTO> chattingRooms;
-}
