+++ src/main/resources/mappers/chatting/chattingRoomMapper.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.freepath.devpath.chatting.query.mapper.ChattingRoomMapper">
+    <select id="selectChattingRooms" resultType="com.freepath.devpath.chatting.query.dto.ChattingRoomDTO">
+        SELECT
+            cr.chatting_room_id,
+            cr.chatting_room_title,
+            cr.user_count
+        FROM
+            CHATTING_JOIN cj
+        JOIN
+            CHATTING_ROOM cr ON cj.chatting_room_id = cr.chatting_room_id
+        WHERE
+            cj.user_id = #{userId}
+            AND cj.chatting_join_status = 'Y';
+    </select>
+</mapper>
